+++ src/Model/Player.h
@@ -40,6 +40,7 @@ public:
     bool isAlive() const;
     bool isDying() const;
     void setDying(bool dying);
+    void markDead();
     bool isFullHp() const;
     double hpRatio() const;
     int handCardLimit() const;
@@ -99,6 +100,7 @@ private:
 
     int m_hp = 0;
     bool m_dying = false;
+    bool m_deathNotified = false;
 
     std::vector<Card*> m_handCards;
     std::vector<Card*> m_equipCards;